requirements.txt
asgiref==3.5.2
Babel==2.11.0
Django==4.1.3
django-money==3.0.0
django-treenode==0.18.2
Pillow==9.3.0
py-moneyed==2.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
